requirement.txt
fastapi
uvicorn
playwright
beautifulsoup4
python-dotenv
openai
firebase-admin
